+++ src/components/projects/ProjectItem.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+const ProjectItem = ({
+    name,
+    icon,
+    projectLink,
+    githubLink,
+    description,
+    imgUrl
+}) => {
+    const divStyle = {
+        backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.7)), url(${imgUrl})`,
+        backgroundSize: `120%`,
+        backgroundPosition: 'center',
+        backgroundRepeat: 'no-repeat'
+    };
+    return (
+        <a
+            href={projectLink}
+            target='_blank'
+            rel='noopener noreferrer'
+            className='project-item'
+            style={divStyle}
+        >
+            <div className='project-item-content'>
+                <p>{name}</p>
+                <FontAwesomeIcon icon={icon} className='my' />
+                <p className='project-description'>{description}</p>
+                <a href={githubLink} className='btn my-1'>
+                    View the code
+                </a>
+            </div>
+        </a>
+    );
+};
+
+export default ProjectItem;