+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
-## Unreleased: feature3/projects-page <> YYYY-MM-DD
+## Unreleased:
+
+---
+
+## feature3/projects-page <> 2021-05-26
 
 ### Added
 
----
+-   Projects & ProjectItem components
+-   Project page setup and routing
 
 ## feature2/contact-page <> 2021-05-26